+++ src/main/java/katecam/luvicookie/ditto/domain/member/application/MemberService.java
@@ -4,7 +4,7 @@ import katecam.luvicookie.ditto.domain.file.application.AwsFileService;
 import katecam.luvicookie.ditto.domain.member.dao.MemberRepository;
 import katecam.luvicookie.ditto.domain.member.domain.Member;
 import katecam.luvicookie.ditto.domain.member.dto.request.MemberCreateRequest;
-import katecam.luvicookie.ditto.domain.member.dto.response.MemberUpdateRequest;
+import katecam.luvicookie.ditto.domain.member.dto.request.MemberUpdateRequest;
 import katecam.luvicookie.ditto.global.error.ErrorCode;
 import katecam.luvicookie.ditto.global.error.GlobalException;
 import lombok.RequiredArgsConstructor;
